+++ __tests__/server/mixins.js
@@ -6,7 +6,7 @@ const mixins = require('../../src/server/mixins')
 describe('mixins', () => {
   let db
 
-  before(() => {
+  beforeAll(() => {
     _.mixin(lodashId)
     _.mixin(mixins)
   })
@@ -25,32 +25,38 @@ describe('mixins', () => {
   })
 
   describe('getRemovable', () => {
-    it('should return removable documents', () => {
+    test('should return removable documents', () => {
       const expected = [
         { name: 'comments', id: 2 },
         { name: 'comments', id: 3 }
       ]
 
-      assert.deepEqual(_.getRemovable(db, { foreignKeySuffix: 'Id' }), expected)
+      assert.deepStrictEqual(
+        _.getRemovable(db, { foreignKeySuffix: 'Id' }),
+        expected
+      )
     })
 
-    it('should support custom foreignKeySuffix', () => {
+    test('should support custom foreignKeySuffix', () => {
       const expected = [
         { name: 'comments', id: 2 },
         { name: 'comments', id: 3 }
       ]
 
-      assert.deepEqual(_.getRemovable(db, { foreignKeySuffix: 'Id' }), expected)
+      assert.deepStrictEqual(
+        _.getRemovable(db, { foreignKeySuffix: 'Id' }),
+        expected
+      )
     })
   })
 
   describe('createId', () => {
-    it('should return a new id', () => {
-      assert.equal(_.createId(db.comments), 4)
+    test('should return a new id', () => {
+      assert.strictEqual(_.createId(db.comments), 4)
     })
 
-    it('should return a new uuid', () => {
-      assert.notEqual(_.createId(db.photos), 3)
+    test('should return a new uuid', () => {
+      assert.notStrictEqual(_.createId(db.photos), 3)
     })
   })
 })