+++ src/cli/bin.js
@@ -1,1 +1,3 @@
+#!/usr/bin/env node
+require('please-upgrade-node')(require('../../package.json'))
 require('./')()