+++ CHANGELOG.md
@@ -1,8 +1,27 @@
 # Change Log
 
-## Unreleased
+## 0.14.2 - 2018-12-26
+
+* Fix `db.getState is not a function`
+
+## 0.14.1 - 2018-12-25
+
+* Show error message if port is already used
+* Upgrade to [lowdb](https://github.com/typicode/lowdb) `1.0`
+
+## 0.14.0 - 2018-06-09
+
+* Listen to `localhost` by default, instead of `0.0.0.0`
+
+## 0.13.0 - 2018-05-30
+
+* Bundle all index page assets so that you access it without network connection
+* Drop Node 4 support
+
+## 0.12.2 - 2018-04-26
 
 * Add `_delay` query parameter
+* Upgrade `please-upgrade-node` dependency
 
 ## 0.12.1 - 2017-11-02